Marge InitialScreen to AuthScreen
