+++ screens/AuthScreen.styled.ts
